server/demo/requirements.txt
aioredis==1.3.1
asgi-redis==1.4.3
asgiref==3.1.2
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.4.3
Automat==20.2.0
certifi==2020.4.5.1
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.9.2
daphne==2.5.0
defusedxml==0.6.0
Django==3.0.6
django-allauth==0.41.0
django-redis==4.12.1
django-rest-auth==0.9.5
django-user-agents==0.4.0
djangorestframework==3.11.0
hiredis==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
Markdown==3.2.2
msgpack==0.6.2
msgpack-python==0.5.6
oauthlib==3.1.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==19.1.0
python-memcached==1.59
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
redis==2.10.6
requests==2.23.0
requests-oauthlib==1.3.0
service-identity==18.1.0
six==1.15.0
sqlparse==0.3.1
Twisted==20.3.0
txaio==20.4.1
ua-parser==0.10.0
urllib3==1.25.9
user-agents==2.1
zipp==3.1.0
zope.interface==5.1.0
